shader bugs

+++ shaders/ui.vert
@@ -1,7 +1,7 @@
 #version 330 core
 layout (location = 0) in vec2 pos;
 layout (location = 1) in vec3 col;
-layout (location = 2) in float floatFlag;
+layout (location = 2) in int floatFlag;
 
 out vec3 color;
 flat out int textureFlag;
